Rename languages dir #12
