+++ frontend/mobile_app_react_native/constants/languages/EN/menu.ts
